baptist_health/timecard_management_agent/deployment/requirements.txt
# Deployment script dependencies
google-cloud-aiplatform[agent_engines]>=1.93.0
google-adk>=1.0.0
google-genai>=1.5.0,<2.0.0
pydantic>=2.10.6,<3.0.0
absl-py>=2.2.1,<3.0.0
python-dotenv>=1.0.0
google-cloud-firestore>=2.18.0
tabulate>=0.9.0
cloudpickle>=3.1.1
